feat: 用户注册｜登录｜信息获取

+++ model/favorite.go
@@ -2,6 +2,6 @@ package model
 
 type Favorite struct {
 	Common
-	UserID  int `json:"user_id"`  // 点赞的用户ID
-	VideoID int `json:"video_id"` // 被点赞的视频ID
+	UserID  uint64 `json:"user_id"`  // 点赞的用户ID
+	VideoID uint64 `json:"video_id"` // 被点赞的视频ID
 }
