+++ pkg/serializer/comment.go
@@ -2,7 +2,7 @@ package serializer
 
 // Comment 评论信息
 type Comment struct {
-	ID         int    `json:"id"`          // 评论id
+	ID         uint64 `json:"id"`          // 评论id
 	User       User   `json:"user"`        // 评论用户
 	Content    string `json:"content"`     // 评论内容
 	CreateDate string `json:"create_date"` // 评论时间
